requirements.txt
django
python-dotenv
pandas
django-plotly-dash
celery
googlemaps
